Check feature flags and enable them if relevant during startup
... just before we verify the Mnesia schema integrity.

The code first checks if the local node is compatible with the remove
nodes (i.e. remote enabled flags are supported locally). If it is not
the case, the startup abort with an error to prevent any damage.

Once the check is passed, we enable locally all feature flags already
enabled remotely. This is shorter code path than the regular `enable()`
function because we don't want or need to mess with remote nodes at this
point: we only need to run the migration function and mark the feature
flag as enabled on the local node.